requirements.txt
neo4j-graphrag[ollama]
vero-eval
ollama
PyPDF2
python-dotenv
fastapi
uvicorn
pydantic
httpx
numpy
matplotlib
plotly
pandas
redis
pinecone-client
python-multipart
nltk
mcp>=0.1.0
aiofiles>=23.0.0
